+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 * Fix velocity, I used the wrong units when I set the default movement velocity. Still need to figure out the caching
   issue and forward the authoritative updates from the server.
 * Add connection mapper for movement events.
+* Send authoritative position and velocity updates from server to client.
 
 #### 20 January 2024